+++ sds_data_manager/lambda_code/efs_lambda/lambda_function.py
@@ -7,11 +7,11 @@ import boto3
 # Define the paths
 mount_path = os.environ.get("EFS_MOUNT_PATH")
 
-attitude_symlink_path = f"{mount_path}/latest_attitude_kernel.ah.a"
-ephemeris_symlink_path = f"{mount_path}/latest_ephemeris_kernel.bsp"
+attitude_symlink_path = Path(f"{mount_path}/latest_attitude_kernel.ah.a")
+ephemeris_symlink_path = Path(f"{mount_path}/latest_ephemeris_kernel.bsp")
 
 
-def create_symlink(source_path: str, destination_path: str) -> None:
+def create_symlink(source_path: Path, destination_path: Path) -> None:
     """Create a symlink from source_path to destination_path
 
     Parameters
@@ -22,14 +22,11 @@ def create_symlink(source_path: str, destination_path: str) -> None:
         Destination path of the symlink
     """
 
-    # Check if the symlink exists
-    if Path(destination_path).is_symlink():
-        # Remove the old symlink
-        Path(destination_path).unlink(missing_ok=True)
+    # Remove the old symlink
+    destination_path.unlink(missing_ok=True)
 
     # Create a new symlink pointing to the new file
-    dest_symlink = Path(destination_path)
-    dest_symlink.symlink_to(source_path)
+    destination_path.symlink_to(source_path)
 
 
 def write_data_to_efs(s3_key: str, s3_bucket: str):
@@ -47,23 +44,22 @@ def write_data_to_efs(s3_key: str, s3_bucket: str):
     s3_client = boto3.client("s3")
 
     # Download the file to the mount directory. Eg. /mnt/efs
-    destination_path = f"{mount_path}/" + os.path.basename(s3_key)
+    download_path = Path(f"{mount_path}/{filename}")
 
     try:
         # Download the file from S3
-        s3_client.download_file(s3_bucket, s3_key, destination_path)
-        logging.debug(f"File downloaded: {destination_path}")
+        s3_client.download_file(s3_bucket, s3_key, download_path)
+        logging.debug(f"File downloaded: {download_path}")
     except Exception as e:
         logging.info(f"Error downloading file: {e!s}")
 
     logging.debug("After downloading file: %s", os.listdir(mount_path))
 
-    source_path = os.path.join(mount_path, filename)
     # Attitude naming convention is this:
     # imap_yyyy_doy_yyyy_doy_##.ah.bc and
     # imap_yyyy_doy_yyyy_doy_##.ah.a
     if filename.endswith(".ah.a") or filename.endswith(".ah.bc"):
-        create_symlink(source_path, attitude_symlink_path)
+        create_symlink(download_path, attitude_symlink_path)
 
     # Ephemeris naming convention is this:
     # Eg. imap_nom_yyyymmdd_yyyymmdd_v##.bsp
@@ -73,7 +69,7 @@ def write_data_to_efs(s3_key: str, s3_bucket: str):
     #   imap_burn_yyyymmdd_yyyymmdd_v##.bsp
     #   imap_pred_yyyymmdd_yyyymmdd_v##.bsp
     elif filename.startswith("imap_nom") and filename.endswith(".bsp"):
-        create_symlink(source_path, ephemeris_symlink_path)
+        create_symlink(download_path, ephemeris_symlink_path)
 
 
 def lambda_handler(event, context):
